Synced Rotation & Synced Animation
With the SyncedRotation and SyncedAnimation scripts, the GameObject will rotate and animate at a speed dependent on the audio's duration.
For example, if an object with SyncedRotation is set to 1 rotation per loop, then it will complete a full rotation in the time it takes for the audio file to finish playing.