Setup Audio Syncing (#35)
* Setup the Conductor class

Keeps track of the song attached to its GameObject

* Updated Conductor Class

Now keeps track of how many times the song has been repeated
Note that for the class to work, you must input the song's BPM and beats per loop *manually*

* Synced Rotation & Synced Animation

With the SyncedRotation and SyncedAnimation scripts, the GameObject will rotate and animate at a speed dependent on the audio's duration.
For example, if an object with SyncedRotation is set to 1 rotation per loop, then it will complete a full rotation in the time it takes for the audio file to finish playing.